fix: get report view to render dynamically in frontend
