+++ src/renderer/components/MapFilter/ReportView/ReportView.js
@@ -5,13 +5,28 @@ import { makeStyles } from '@material-ui/core/styles'
 import ViewWrapper, { type CommonViewProps } from '../ViewWrapper'
 import Toolbar from '../internal/Toolbar'
 import PrintButton from './PrintButton'
+import { FormattedTime, IntlProvider } from 'react-intl'
 import HideFieldsButton from './HideFieldsButton'
 import { fieldKeyToLabel } from '../utils/strings'
 import getStats from '../stats'
-import api from '../../../new-api'
+import {
+  Page,
+  Text,
+  View,
+  Image,
+  Document,
+  StyleSheet
+} from '@react-pdf/dom'
+const { isEmptyValue } = require('../utils/helpers')
+const { get } = require('../utils/get_set')
+const FormattedFieldname = require('../internal/FormattedFieldname')
+const FormattedValue = require('../internal/FormattedValue')
+const FormattedLocation = require('../internal/FormattedLocation')
 
 import type { Observation } from 'mapeo-schema'
 import type { PresetWithAdditionalFields, FieldState, Field } from '../types'
+
+const PdfContext = React.createContext(false)
 
 type Props = {
   ...$Exact<CommonViewProps>
@@ -57,6 +72,8 @@ const ReportView = ({
       })
   })
 
+  const cx = useStyles()
+
   return (
     <ViewWrapper
       observations={observations}
@@ -88,14 +105,19 @@ const ReportView = ({
           return obs
         })
 
-        // ReportPageContent defined below...
-        return <ReportPageContent
-          mapStyle={mapStyle}
-          mapboxAccessToken={mapboxAccessToken}
-          fieldState={fieldState}
-          onFieldStateUpdate={setFieldState}
-          observations={observations}
-        />
+        return <div className={cx.root}>
+          <Toolbar>
+            <HideFieldsButton
+              fieldState={fieldState}
+              onFieldStateUpdate={setFieldState}
+            />
+          </Toolbar>
+          <ReportPageContent
+            mapStyle={mapStyle}
+            mapboxAccessToken={mapboxAccessToken}
+            observations={observations}
+          />
+      </div>
       }}
     </ViewWrapper>
   )
@@ -104,43 +126,26 @@ const ReportView = ({
 const ReportPageContent = ({
   observations,
   fieldState,
-  onFieldStateUpdate,
   mapboxAccessToken,
   mapStyle
 }) => {
-  const cx = useStyles()
-  const [reportId, setReportId] = useState()
+  return <Document>
+    <Page size="A4" style={styles.page} wrap>
+      <Text render={({ pageNumber, totalPages }) => (
+        `${pageNumber} / ${totalPages}`
+      )} fixed />
 
-  useEffect(() => {
-    var promise = api.createReport({
-      observations,
-      fieldState,
-      mapboxAccessToken,
-      mapStyle
-    })
-    promise.then((_reportId) => {
-      console.log('got back reportId:', reportId)
-      if (_reportId) setReportId(_reportId)
-      // TODO: what to do when reportId is undefined?
-    })
-  }, [])
-
-  var link = `/report/${reportId}`
-
-  return (
-    <div className={cx.root}>
-      <Toolbar>
-        {reportId && <a href={link} download>
-          <PrintButton />
-        </a>}
-        <HideFieldsButton
-          fieldState={fieldState}
-          onFieldStateUpdate={onFieldStateUpdate}
-        />
-      </Toolbar>
-      <iframe width="100%" height="100%" src={link} />
-    </div>
-  )
+      <View render={({ pageNumber }) => {
+        const observation = observations[pageNumber]
+        console.log('rendering', pageNumber, observation)
+        return observation && <FeaturePage
+          key={observation.id}
+          observation={observation}
+          />
+        }
+      } />
+    </Page>
+  </Document>
 }
 
 function hiddenFieldsFilter (fieldState: FieldState) {
@@ -168,3 +173,169 @@ const useStyles = makeStyles(theme => ({
     flexDirection: 'column'
   }
 }))
+
+const FeaturePage = ({
+  observation
+}) => {
+  const coords =
+    typeof observation.lon === 'number' && typeof observation.lat === 'number'
+      ? {
+          longitude: observation.lon,
+          latitude: observation.lat
+        }
+      : undefined
+  const createdAt =
+    typeof observation.created_at === 'string'
+      ? new Date(observation.created_at)
+      : undefined
+
+  const preset = observation.preset
+  const fields = preset.fields.concat(preset.additionalFields)
+  const tags = observation.tags || {}
+  const note = tags.note || tags.notes
+
+  return (
+    <View>
+      <View style={styles.header} fixed />
+      <View style={styles.pageContent}>
+        <View style={styles.columnLeft}>
+          <Text style={styles.presetName}>{preset.name || 'Observation'}</Text>
+          {createdAt && (
+            <Text style={styles.createdAt}>
+              <Text style={styles.createdAtLabel}>Registrado: </Text>
+              <FormattedTime
+                key="time"
+                value={createdAt}
+                year="numeric"
+                month="long"
+                day="2-digit"
+              />
+            </Text>
+          )}
+          {coords && (
+            <Text style={styles.location}>
+              <Text style={styles.locationLabel}>Ubicación: </Text>
+              <FormattedLocation {...coords} />
+            </Text>
+          )}
+          {note &&
+            note.split('\n').map((para, idx) => (
+              <Text key={idx} style={styles.description}>
+                {para}
+              </Text>
+            ))}
+          <Text style={styles.details}>Detalles</Text>
+          {fields.map(field => {
+            const value = get(tags, field.key)
+            if (isEmptyValue(value)) return null
+            return (
+              <View key={field.id} style={styles.field} wrap={false}>
+                <Text style={styles.fieldLabel}>
+                  <FormattedFieldname field='test' />
+                </Text>
+                <Text style={styles.fieldValue}>
+                  <FormattedValue field='test' value={value} />
+                </Text>
+              </View>
+            )
+          })}
+        </View>
+        <View style={styles.columnRight}>
+          {observation.attachments && observation.attachments.slice(0, 4).map((att, i) => {
+            return att.media && <Image
+              src={att.media.src}
+              key={i}
+              style={styles.image}
+              wrap={false}
+            />
+          }
+          )}
+        </View>
+      </View>
+      <View style={styles.footer} fixed />
+    </View>
+  )
+}
+
+const styles = StyleSheet.create({
+  page: {
+    backgroundColor: 'white',
+    paddingTop: 120,
+    paddingBottom: 35,
+    paddingHorizontal: 35,
+    flexDirection: 'row'
+  },
+  pageContent: {
+    flex: 1,
+    flexDirection: 'row'
+  },
+  columnLeft: {
+    flex: 2,
+    paddingRight: 12,
+    lineHeight: 1.2
+  },
+  columnRight: {
+    // backgroundColor: 'aqua',
+    flex: 1
+  },
+  presetName: {
+    fontWeight: 700
+  },
+  createdAt: {
+    fontSize: 12
+  },
+  createdAtLabel: {
+    fontSize: 12,
+    color: 'grey'
+  },
+  location: {
+    fontSize: 12,
+    marginBottom: 6
+  },
+  locationLabel: {
+    fontSize: 12,
+    color: 'grey'
+  },
+  map: {
+    height: '60mm',
+    borderStyle: 'solid',
+    borderWidth: 1,
+    borderColor: 'black',
+    marginBottom: 12,
+    backgroundColor: '#8E918B'
+  },
+  image: {
+    height: '40mm',
+    borderStyle: 'solid',
+    borderWidth: 1,
+    borderColor: 'black',
+    marginBottom: 10,
+    backgroundColor: '#C8D8E3'
+  },
+  description: {
+    marginBottom: 6,
+    fontSize: 12
+  },
+  details: {
+    fontWeight: 'bold',
+    fontSize: 14,
+    marginBottom: 3,
+    marginTop: 12
+  },
+  field: {
+    marginBottom: 6
+  },
+  fieldLabel: {
+    fontSize: 9,
+    marginBottom: 1,
+    color: '#333333'
+  },
+  fieldValue: {
+    fontSize: 12
+  },
+  header: {
+  },
+  footer: {
+  }
+
+})
